refactoring the llamacpp client
